+++ price-feeder/CHANGELOG.md
@@ -1,4 +1,5 @@
 <!-- markdownlint-disable MD013 MD024 -->
+<!-- markdown-link-check-disable -->
 
 <!--
 Changelog Guiding Principles:
@@ -56,6 +57,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [#1032](https://github.com/umee-network/umee/pull/1032) Update the accepted tvwap period from 3 minutes to 5 minutes.
 - [#978](https://github.com/umee-network/umee/pull/978) Cleanup the oracle package by moving deviation & conversion logic.
 - [#1175](https://github.com/umee-network/umee/pull/1175) Add type ProviderName.
+- [#1255](https://github.com/umee-network/umee/pull/1255) Move TickerPrice and CandlePrice to types package
 
 ### Features